KDFS-windows-unicode-support-20080511
LICENSE MIT

In NTCreateX a zero length name is permitted.  This can occur when
opening the root directory of a share.

Change cm_NameEntries() to accept a size_t* instead of long* since
that permits better type safety.


(cherry picked from commit 9d28d95151418189e8004734eca0edd498edd46b)

+++ src/WINNT/afsd/cm_dir.c
@@ -121,7 +121,7 @@ cm_DirFreeBlobs(cm_dirOp_t * op, int firstblob, int nblobs);
  * the specified name.
  */
 long 
-cm_NameEntries(char *namep, long *lenp)
+cm_NameEntries(char *namep, size_t *lenp)
 {
     long i;
